svelte(phase 3): migrate RightSidebar chrome into RightSidebar.svelte
Absorb right-sidebar.js (scratchpad load/save, resize drag, tab
switching, expand/collapse) into RightSidebar.svelte's onMount. The
component exposes its controls on window.__piRightSidebar so
session-ui-runner and the annotation/keyboard wiring keep working.
Delete right-sidebar.js and port its tests to RightSidebar.test.js.
